# Conflicts:
#	.idea/deploymentTargetSelector.xml
#	.idea/misc.xml
#	app/build.gradle.kts
#	app/src/main/java/com/example/quiz/data/AuthRepositoryImpl.kt
#	app/src/main/java/com/example/quiz/data/QuizRepository.kt
#	app/src/main/java/com/example/quiz/di/AppModule.kt
#	app/src/main/java/com/example/quiz/navigation/QuizNavHost.kt
#	app/src/main/java/com/example/quiz/ui/feature/auth/AuthViewModel.kt
#	app/src/main/java/com/example/quiz/ui/feature/home/HomeScreen.kt
#	app/src/main/java/com/example/quiz/ui/feature/home/HomeViewModel.kt
#	app/src/main/java/com/example/quiz/ui/feature/login/LoginScreen.kt
#	app/src/main/java/com/example/quiz/ui/feature/quiz/QuizScreen.kt
#	app/src/main/java/com/example/quiz/ui/feature/singup/SingupScreen.kt
#	app/src/main/java/com/example/quiz/ui/feature/theme/Theme.kt
#	app/src/main/java/com/example/quiz/ui/feature/theme/Type.kt

+++ app/src/main/java/com/example/quiz/navigation/QuizNavHost.kt
@@ -34,7 +34,7 @@ object QuizRoute
 data class QuizResultRoute(
     val totalQuestions: Int,
     val correctAnswers: Int,
-    val scorePercentage: Double,
+    val scorePercentage: Float,
     val totalTimeSeconds: Long
 )
 
@@ -106,7 +106,7 @@ fun AppNavigation() {
                         QuizResultRoute(
                             totalQuestions = total,
                             correctAnswers = correct,
-                            scorePercentage = percentage,
+                            scorePercentage = percentage.toFloat(),
                             totalTimeSeconds = time
                         )
                     ) {
@@ -122,7 +122,7 @@ fun AppNavigation() {
             QuizResultScreen(
                 totalQuestions = route.totalQuestions,
                 correctAnswers = route.correctAnswers,
-                scorePercentage = route.scorePercentage,
+                scorePercentage = route.scorePercentage.toDouble(),
                 totalTimeSeconds = route.totalTimeSeconds,
                 onPlayAgain = {
                     navController.navigate(QuizRoute) {
